TCIC2-3 Reorganized the database module a bit which required some changes in other modules to keep them up to date
