+++ database/src/login/mod.rs
@@ -1,0 +1,3 @@
+
+pub mod bot;
+pub mod user;
